Execution folder renamed to stin. Code folder deleted and code files are placedin stin. In setup.py file, packaged line is changed.
