+++ stin/input.yaml
